requirements.txt
datasets
torch
transformers
openai
python-dotenv
sentence-transformers